confirm removing connection from start page
git-svn-id: svn+ssh://svn.code.sf.net/p/webissues/code/trunk/client@1391 c8fa975d-6fd4-4e76-b3b8-1a99b2ee3549

+++ src/views/startview.cpp
@@ -40,6 +40,7 @@
 #include <QTreeWidget>
 #include <QMenu>
 #include <QAction>
+#include <QMessageBox>
 
 StartView::StartView( QObject* parent, QWidget* parentWidget ) : View( parent ),
     m_batch( NULL ),
@@ -184,8 +185,12 @@ void StartView::removeConnection()
 {
     Bookmark bookmark = selectedBookmark();
     if ( !bookmark.url().isEmpty() ) {
-        application->bookmarksStore()->deleteBookmark( bookmark );
-        delete m_list->selectedItems().first();
+        if ( QMessageBox::warning( mainWidget(), tr( "Warning" ),
+             tr( "Are you sure you want to remove the selected connection?" ),
+             QMessageBox::Ok | QMessageBox::Cancel ) == QMessageBox::Ok ) {
+            application->bookmarksStore()->deleteBookmark( bookmark );
+            delete m_list->selectedItems().first();
+        }
     }
 }
